+++ includes/light.h
@@ -3,10 +3,10 @@
 
 # include "raytrace.h"
 
-static inline t_color	diffuse_light(t_light *l, t_hit *hit, double val[2])
+static inline t_color	diffuse_light(t_light *l, t_hit *hit, float val[2])
 {
 	t_color	color;
-	double	k;
+	float	k;
 
 	k = hit->object->material->diffuse;
 	color = v4_multi(val[0] * k, hit->object->color);
@@ -18,20 +18,20 @@ static inline t_color	spec_light(t_light *light, t_hit *hit, t_v3 light_v)
 {
 	t_v3	reflect_v;
 	t_v3	v;
-	double	reflect_dot_eye;
-	double	factor;
+	float	reflect_dot_eye;
+	float	factor;
 
-	v = v3_multi(-1, light_v);
+	v = v3_multi(-1.0f, light_v);
 	reflect_v = reflect(&hit->normal, &v);
 	reflect_dot_eye = ft_dot(reflect_v, normalize(hit->ray->o));
 	if (reflect_dot_eye <= 0)
-		return (v4(0, 0, 0, 1));
-	factor = pow(reflect_dot_eye, hit->object->material->shininess);
+		return (v4(0.0f, 0.0f, 0.0f, 1.0f));
+	factor = powf(reflect_dot_eye, hit->object->material->shininess);
 	factor = factor * hit->object->material->specular;
 	return (v4_multi(factor, light->color));
 }
 
-static inline bool	is_shadowed(t_scene *world, t_v3 to_l, t_point p, double d)
+static inline bool	is_shadowed(t_scene *world, t_v3 to_l, t_point p, float d)
 {
 	t_ray	*shadow_ray;
 	t_array	*records;
@@ -55,7 +55,7 @@ t_color	lighting(t_scene *scn, t_hit *hit, t_light *light)
 	t_color	phong[3];
 	t_color	color;
 	t_v3	to_light;
-	double	val[2];
+	float	val[2];
 
 	to_light = v3_sub(light->o, hit->h_point);
 	val[1] = v3_norm_2(to_light);
@@ -64,8 +64,8 @@ t_color	lighting(t_scene *scn, t_hit *hit, t_light *light)
 	if (is_shadowed(scn, to_light, hit->over_p, val[1]))
 		return (phong[0]);
 	val[0] = ft_dot(to_light, hit->normal);
-	phong[1] = v4(0, 0, 0, 1);
-	phong[2] = v4(0, 0, 0, 1);
+	phong[1] = v4(0.0f, 0.0f, 0.0f, 1.0f);
+	phong[2] = v4(0.0f, 0.0f, 0.0f, 1.0f);
 	if (val[0] >= 0)
 	{
 		phong[1] = diffuse_light(light, hit, val);
